feat: Use req.user.id in getProfile
In the 'getProfile' function of the 'authController', the user ID was previously being accessed from 'req.userId'. This has been updated to use 'req.user.id' to maintain consistency with the 'authenticate' middleware, which sets the user information on the 'req.user' object.

The corresponding tests for 'getProfile' have also been updated to reflect this change.